Factor out base results page from results listing page
